services/client/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.4
Flask-Testing==0.8.0
gunicorn==19.8.1